pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lora-trainer"
version = "0.1.0"
description = "Minimal yet practical LoRA trainer with CLI-first reproducibility."
authors = [
    {name = "lora-trainer contributors"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["lora", "stable-diffusion", "training", "diffusion-models"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.25.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
xformers = [
    "xformers>=0.0.22",
]

[project.scripts]
lora-trainer = "lora_trainer.cli:main"
lora-colab-agent = "lora_trainer.colab_agent:main"

[project.urls]
Homepage = "https://github.com/yourusername/lora-trainer"
Documentation = "https://lora-trainer.readthedocs.io"
Repository = "https://github.com/yourusername/lora-trainer"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501", "N812"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
